Migrate signature scheme into separate module
